Add logic for implement error in javascript

+++ app/Http/Controllers/StaplesFoxController.php
@@ -28,7 +28,7 @@ class StaplesFoxController extends ApiFoxController
         ):object|string
     {
      
-        
+        //Validate URL structure
         if($validateurl === true && $validatehttps === true){
 
             //Get Http response status
@@ -64,7 +64,7 @@ class StaplesFoxController extends ApiFoxController
                     $get_title = implode("", $this->filterData($product_data, 'title'));
                     //$get_price = intval(str_replace('$', '', implode($this->filterData($product_data, 'price'))));
                     $get_price = floatval(str_replace('$', '', implode($this->filterData($product_data, 'price'))));
-                    $get_total_of_starts = intval(implode("", $this->filterData($product_data, 'stars')));
+                    $get_total_of_starts = floatval(implode("", $this->filterData($product_data, 'stars')));
                     $get_total_of_reviews = intval(str_replace(' Review', '', implode("", $this->filterData($product_data, 'reviews'))));
 
                     //List of Clean Data
